pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bot-bitirouvchi"
version = "0.1.0"
description = "Telegram survey bot"
authors = [{name = "Student"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.3.0",
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
    "pandas>=2.1.0",
    "openpyxl>=3.1.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]

